+++ app/src/main/java/com/example/ishopping/data/source/local/BookmarkItemDao.kt
@@ -1,9 +1,11 @@
-package com.example.ishopping.data.model
+package com.example.ishopping.data.source.local
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import com.example.ishopping.data.model.BookmarkItem
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface BookmarkItemDao {
@@ -14,5 +16,8 @@ interface BookmarkItemDao {
     fun delete(bookmarkItem: BookmarkItem)
 
     @Query("SELECT * FROM bookmark_items")
-    fun getAll(): List<BookmarkItem>
+    fun getAll(): Flow<List<BookmarkItem>>
+
+    @Query("DELETE FROM bookmark_items")
+    suspend fun clearTable()
 }
